fix(input): multi place builder patches
