+++ plugin/src/camera/patch/builder/builder_raycast_patch.cs
@@ -5,7 +5,7 @@ using HarmonyLib;
 using Plugin.Input;
 using UnityEngine;
 
-namespace TechtonicaVR.VRCamera.Patch;
+namespace TechtonicaVR.VRCamera.Patch.Builder;
 
 [HarmonyPatch]
 public class BuilderRaycastPatch
